+++ src/cli-branch.ts
@@ -4,6 +4,12 @@ import { CLI_BRANCH } from './constants';
 
 type Config = ExcludeCommandType<CliBranch>;
 
+/**
+ * A factory function for creating command branches
+ * @remarks
+ * @param config
+ * @returns Returns the newly-created `CliBranch` object
+ */
 export function CliBranch(config: Config): CliBranch {
   const branch: CliBranch = {
     ...config,