+++ src/index.ts
@@ -2,24 +2,22 @@
 export { CliBranch } from './cli-branch';
 export { CliLeaf } from './cli-leaf';
 
+// CliArgParser factories
+export { CliFlagArgParser } from './arg-parsers/cli-flag-arg-parser';
+export { CliJsonArgParser } from './arg-parsers/cli-json-arg-parser';
+export { CliNumberArgParser } from './arg-parsers/cli-number-arg-parser';
+export { CliNumberArrayArgParser } from './arg-parsers/cli-number-array-arg-parser';
+export { CliOneOfArgParser } from './arg-parsers/cli-one-of-arg-parser';
+export { CliStringArgParser } from './arg-parsers/cli-string-arg-parser';
+export { CliStringArrayArgParser } from './arg-parsers/cli-string-array-arg-parser';
+
 // Command-line interface runner
 export { runCliAndExit } from './run-cli-and-exit';
-
-// CliArgParser factories:
-export { CliFlagArgParser } from './arg-parser-factories/cli-flag-arg-parser';
-export { CliJsonArgParser } from './arg-parser-factories/cli-json-arg-parser';
-export { CliNumberArgParser } from './arg-parser-factories/cli-number-arg-parser';
-export { CliNumberArrayArgParser } from './arg-parser-factories/cli-number-array-arg-parser';
-export { CliOneOfArgParser } from './arg-parser-factories/cli-one-of-arg-parser';
-export { CliStringArgParser } from './arg-parser-factories/cli-string-arg-parser';
-export { CliStringArrayArgParser } from './arg-parser-factories/cli-string-array-arg-parser';
+export { RunCli, CliEnhancer } from './run-cli';
 
 // Error constructors
 export { CliTerseError, CLI_TERSE_ERROR } from './cli-terse-error';
 export { CliUsageError, CLI_USAGE_ERROR } from './cli-usage-error';
 
-// CliArgRunner
-export { RunCli as CliArgRunner, CliEnhancer } from './run-cli';
-
-// For custom arg types
+// For custom arg parsers
 export { CliArgParser } from './types';