+++ changelog.md
@@ -1,10 +1,11 @@
 # @carnesen/cli change log
 
-## [Unreleased]
+## [Unreleased - 0.5.0]
 ### Breaking
 Refactor:
 - "input" --> "arg parser" 
 - "getValue" --> "parse"
+- "ArgvInterface" --> "RunCli"
 ## 0.4.0
 ### Breaking
 Numerous tweaks to make the API more expressive and grepable, e.g. CliLeaf instead of just Leaf. The docs are still a work-in-progress but the TypeScript types are spot-on. Let them be your guide!